Move: 아이콘 파일 assets 내 icons 폴더로 이동
