+++ src/components/Icon/Icon.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import iconSvg from '../../assets/icon.svg';
+import iconSvg from '../../assets/icons/icon.svg';
 
 export default function Icon({ id, fill, ...props }) {
   return (
